Sema: Move some code from TypeCheckDecl.cpp to TypeCheckEffects.cpp
